+++ tests/ext/aws/test_providers.py
@@ -4,7 +4,7 @@ import pytest
 from asynctest import CoroutineMock
 from botocore.exceptions import BotoCoreError, ClientError
 
-from loafer.exceptions import ProviderError
+from loafer.exceptions import ProviderError, ProviderRuntimeError
 from loafer.ext.aws.providers import SQSProvider
 
 
@@ -107,3 +107,13 @@ def test_stop():
     provider._client_stop = CoroutineMock()
     provider.stop()
     provider._client_stop.assert_awaited()
+
+
+@pytest.mark.asyncio
+def test_reraise_runtime_error_as_sqs_provider_runtime_error(mock_boto_session_sqs, boto_client_sqs):
+    with mock_boto_session_sqs:
+        provider = SQSProvider('queue-name')
+        boto_client_sqs.close.side_effect = RuntimeError()
+
+        with pytest.raises(ProviderRuntimeError):
+            provider.stop()